feat: change module sharing strategy
BREAKING CHANGE: Previously, we used to "rekey" all shared packages used in a host in order to prevent eager consumption issues. However, this caused unforeseen issues when trying to share a singleton package, as the package would end up being bundled multiple times per page.

As a result, we have had to stop rekeying shared modules in userland and only do so on internal Next packages themselves.

If you need to dangerously share a package using the old method, you can do so by using the following code:

                 const shared = {
                   fakeLodash: {
                     import: "lodash",
                     shareKey: "lodash",
                   }
                 }

Please note that this method is now considered dangerous and should be used with caution.