feat: Quantum Modules
Shared modules that can coexist as both eager and lazy/external within on module graph.

To the "host" eager sharing keeps next/react and so on in main.

to the remote, this eager module reference is removed, along with the "provide shared" module reference which is responsible for registering that dependency on share scope.

Doing so makes the remote container aleays consume react, but never register it nor try to load its fallback/eager reference which is removed.

In the future, Quantum modules could create ConsumeSharedModule dependencies to actually offer lazy backups to the remote, but eager provide in the host.