feat: Quantum Modules (#872)
* feat: Quantum Modules

Shared modules that can coexist as both eager and lazy/external within on module graph.

To the "host" eager sharing keeps next/react and so on in main.

to the remote, this eager module reference is removed, along with the "provide shared" module reference which is responsible for registering that dependency on share scope.

Doing so makes the remote container aleays consume react, but never register it nor try to load its fallback/eager reference which is removed.

In the future, Quantum modules could create ConsumeSharedModule dependencies to actually offer lazy backups to the remote, but eager provide in the host.

* fix: implement promise queue cleanup

* feat: RemoveEagerModulesFromRuntimePlugin

* Refactor InvertedContainerPlugin to use runtimeTemplate and add fancyTemplate

This commit refactors the InvertedContainerPlugin to use runtimeTemplate instead of javascript.JavascriptModulesPlugin.getCompilationHooks(). It also adds a fancyTemplate to improve the readability of the code. The changes made to the code include:

### packages/nextjs-mf/src/plugins/container/InvertedContainerPlugin.ts

- Refactor InvertedContainerPlugin to use runtimeTemplate instead of javascript.JavascriptModulesPlugin.getCompilationHooks()
- Add fancyTemplate to improve code readability