fix: Improved Share Scope Properties and Fixed Production Build Issues (#884)
* chore(nextjs-mf): release version 6.4.1-beta.8

* thing

* fix: fixed production build issues

* fix: Prevent removal of external eager modules

This commit prevents the removal of eager modules that are external, improves the code by removing an unnecessary condition, and refactors test patterns for better readability.

### next-fragments.ts

- Removed unnecessary condition checking 'compiler.options.name === 'client' || true'

- Refactored 'test' patterns for better readability

### RemoveEagerModulesFromRuntimePlugin.ts

- Prevented removal of external eager modules

* chore: update packscript

* Update default share scope properties

This commit updates the DEFAULT_SHARE_SCOPE and DEFAULT_SHARE_SCOPE_BROWSER objects in the internal.ts file. Modifications include setting the eager property for some shared dependencies and using direct reference instead of cloning the DEFAULT_SHARE_SCOPE object.

### packages/nextjs-mf/src/internal.ts

- Set eager property for 'react-dom' to false in DEFAULT_SHARE_SCOPE

- Change import property order for 'react/jsx-runtime' in DEFAULT_SHARE_SCOPE

- Remove JSON clone process and use a direct reference of DEFAULT_SHARE_SCOPE in building DEFAULT_SHARE_SCOPE_BROWSER

- Set eager property to true for 'react', 'react-dom', and 'next/router' in DEFAULT_SHARE_SCOPE_BROWSER

* Fix incorrect string in if condition

This commit fixes an issue in the patchDefaultSharedLoader.ts file where the if condition check was using an incorrect string, which could result in skipping the intended patching process.

### packages/nextjs-mf/src/loaders/patchDefaultSharedLoader.ts

- Update the string in the if condition check from 'placeholderModuleEnsurepat' to 'placeholderModuleEnsure'

* Refactor loadScript and improve error handling

This commit refactors the loadScript function and improves error handling for offline containers. It also fixes type casting issues in the code.

### packages/utilities/src/utils/common.ts

- Fix type casting for remoteGlobal and containerKey

- Enable console.error statement for error logging when catching errors

- Refactor the loadScript function to improve readability and error handling