fix: Runtime module checking (#969)
* fix: dont force named chunkids on client

server builds seem to fall over if chunk ids are numeric, however this is less of a problem when in serverland

* refactor: move delegate module utils to own file

* refactor: point default delegate to origin file

* fix: safety checks for getEagerSharedForChunkId

if no initial chunks exist, runtime modules might not be added. As such, we should check that the runtime module exists before calling it

* fix: safety checks for getEagerSharedForChunkId

if no initial chunks exist, runtime modules might not be added. As such, we should check that the runtime module exists before calling it

* fix: safety checks for getEagerSharedForChunkId

if no initial chunks exist, runtime modules might not be added. As such, we should check that the runtime module exists before calling it