refactor: enhanced global object proxy handling
In this commit, a few significant enhancements have been made to the handling of the global object proxy. Adjustments are made in the handling of properties 'remotes', 'cache', and '_config' - now situated under the '__remote_scope__' property of the global object. Additionally, console log statements have been added to aid in debugging.

### DynamicFilesystemChunkLoadingRuntimeModule.ts

- refactored the structure of the 'remotes' object

- changed federation remotes population method

- added a console log statement for debugging

### FederationModuleInfoRuntimeModule.ts

- created a new ES6 Proxy on '__remote_scope__' to manage 'remotes' and 'cache'

- augmented logging to trace the reading and writing operations on properties

### loadScript.ts

- enhanced error handling in 'executeLoad' function

- adjusted assignment of '_config' and 'remotes' properties in global object

- added debugging console log statements