refactor: Adjustments to 'default-delegate.ts' and 'NextFederationPlugin/index.ts' files, and addition to 'RemotePublicPathRuntimeModule.js'
This commit includes refactoring changes and additions to three files. For 'default-delegate.ts', new code was added to several functions and log statements, while some white space and commented code was removed in multiple locations. In 'NextFederationPlugin/index.ts', debug condition is un-commented, and in 'RemotePublicPathRuntimeModule.js', a license comment was added.

### packages/nextjs-mf/src/default-delegate.ts

- Added new code to 'loadScript' and 'getRuntimeRemotes' functions

- Removed extra white space and commented code

### packages/nextjs-mf/src/plugins/NextFederationPlugin/index.ts

- Un-commented the debug condition

### packages/node/src/plugins/RemotePublicPathRuntimeModule.js

- Added the MIT License comment