feat: add auto public path support for remote modules (node_auto_public_path)
This commit adds support for automatically setting the public path for
remote modules. This feature allows for dynamic path resolution based on
the publicPath value in the outputOptions of the webpack configuration.
If the publicPath value is set to "auto", the path is automatically
resolved, otherwise, the static value of publicPath is used. This change
provides more flexibility in setting the public path for remote modules.