remove usage of createDelegatedModule from nextjs-mf (#node-sync-bytedance)
The createDelegatedModule utility from the nextjs-mf package was commented out,
indicating no longer necessary for the checkout app's next.config.js. Also, a
jest environment was added to the eslint configuration in the node package, and
the generated comments and ts-ignore directives were removed from the test file
for parseRemotes in the NodeFederationPlugin.