feat(ContainerEntryModule): fix TypeScript issue, enhance needBuild method #398a60e
This commit addresses a TypeScript issue where the 'needBuild' property in
'ContainerEntryModule' was not assignable to the same property in the base
type 'Module'. This was due to incompatible types of the 'callback' and
'error' parameters. The fix involved changing the type of 'error' parameter in
the callback and making the 'details' property in 'WebpackError' required.